5.1.0 first step, see ChangeLog
git-svn-id: https://stlport.svn.sourceforge.net/svnroot/stlport/trunk/STLport@1872 01b27279-081b-0410-8cf9-d9d9e0fb0389

+++ test/eh/random_number.cpp
@@ -23,7 +23,7 @@
 # include <stdlib.h>
 #endif
 
-unsigned random_number( unsigned range )
+unsigned random_number( size_t range )
 {
 #if !defined( __SGI_STL )
   if (range == 0) return 0;
